@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -42,7 +42,6 @@ dependencies:
   firebase_storage: ^11.0.14
   file_picker: ^5.2.5
   flutter_signin_button: ^2.0.0
-  flutter_icons: ^1.1.0
   multi_select_flutter: ^4.1.2
   pdf: ^3.8.3
   printing: ^5.9.2
