@@ -4,7 +4,8 @@ class Responsive extends StatelessWidget {
   final Widget mobile;
   final Widget desktop;
 
-  const Responsive({Key key, this.mobile, this.desktop}) : super(key: key);
+  const Responsive({Key? key, required this.mobile, required this.desktop})
+      : super(key: key);
 
   static bool isMobile(BuildContext context) =>
       MediaQuery.of(context).size.width < 600;
