@@ -1,7 +1,6 @@
 import 'package:extensiona_if/components/editor.dart';
 import 'package:extensiona_if/validation/validation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_icons/flutter_icons.dart';
 
 class LoginForm extends StatelessWidget {
   final TextEditingController emailController;
@@ -9,7 +8,10 @@ class LoginForm extends StatelessWidget {
   final GlobalKey<FormState> formKey;
 
   const LoginForm(
-      {Key key, this.emailController, this.passwordController, this.formKey})
+      {Key? key,
+      required this.emailController,
+      required this.passwordController,
+      required this.formKey})
       : super(key: key);
   @override
   Widget build(BuildContext context) {
@@ -35,7 +37,7 @@ class LoginForm extends StatelessWidget {
             controlador: passwordController,
             rotulo: 'Senha',
             dica: 'Informe a sua senha',
-            icon: const Icon(Ionicons.md_key),
+            icon: const Icon(Icons.key_rounded),
             qtdCaracteres: 10,
             verSenha: true,
             confirmPasswordField: false,
