@@ -4,7 +4,7 @@ import 'package:extensiona_if/components/item_demanda.dart';
 import 'package:extensiona_if/theme/app_theme.dart';
 
 class ListaDemanda extends StatelessWidget {
-  const ListaDemanda({Key key}) : super(key: key);
+  const ListaDemanda({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -17,4 +17,3 @@ class ListaDemanda extends StatelessWidget {
     );
   }
 }
-
